+++ SourceCode/Database/dbo/Tables/Property.sql
@@ -4,10 +4,6 @@
     [Value] NVARCHAR (50) NOT NULL,
     CONSTRAINT [PK_Property] PRIMARY KEY CLUSTERED ([Id] ASC)
 );
-
-
-
-
 GO