+++ SourceCode/Services/Extensions/LocoAddressExtensions.cs
@@ -3,11 +3,12 @@
 namespace ModulesRegistry.Services.Extensions;
 public static class LocoAddressExtensions
 {
-    public static int[] AsLocoAdresses(this string? values)
+    public static bool TryParseLocoAdresses(this string? values, out int[] addresses)
     {
-        if (string.IsNullOrWhiteSpace(values)) return Array.Empty<int>();
+        addresses = Array.Empty<int>();
+        if (string.IsNullOrWhiteSpace(values)) return true;
         var itemGroups = values.Split(',');
-        if (itemGroups.Length == 0) return Array.Empty<int>();
+        if (itemGroups.Length == 0) return true;
         var result = new List<int>();
         foreach (var group in itemGroups)
         {
@@ -15,13 +16,25 @@ public static class LocoAddressExtensions
             var trimmedGroup = group.Trim();
             var interval = trimmedGroup.Split('-');
             if (interval.Length == 0) continue;
-            if (interval.Length == 1 && int.TryParse(interval[0], out int address) && IsValidDccAddress(address)) result.Add(address);
-            if (interval.Length == 2 && int.TryParse(interval[0], out int fromAddress) && int.TryParse(interval[1], out int toAddress) && fromAddress < toAddress && IsValidDccAddress(fromAddress) && IsValidDccAddress(toAddress)) result.AddRange(Enumerable.Range(fromAddress, toAddress - fromAddress + 1));
-            if (interval.Length > 2) throw new ArgumentOutOfRangeException(nameof(values), values);
+            if (interval.Length == 1)
+            {
+                if (int.TryParse(interval[0], out int address) && address.IsValidDccAddress())
+                    result.Add(address);
+                else
+                    return false;
+            }
+            if (interval.Length == 2)
+            {
+                if (int.TryParse(interval[0], out int fromAddress) && int.TryParse(interval[1], out int toAddress) && fromAddress < toAddress && fromAddress.IsValidDccAddress() && toAddress.IsValidDccAddress())
+                    result.AddRange(Enumerable.Range(fromAddress, toAddress - fromAddress + 1));
+                else
+                    return false;
+            }
+            if (interval.Length > 2) return false;
         }
-        return result.ToArray();
+        addresses = result.ToArray();
+        return true;
 
-        static bool IsValidDccAddress(int address) => address >= 1 && address <= 9999;
     }
 
     public static string AsCollapsedLocoAdresses(this int[]? adresses)
@@ -29,7 +42,7 @@ public static class LocoAddressExtensions
         if (adresses is null || adresses.Length == 1) return string.Empty;
         var result = new StringBuilder(200);
         int intervalStartIndex = -1;
-        var orderedAdresses = adresses.OrderBy(a => a).ToArray();
+        var orderedAdresses = adresses.Where(a => a.IsValidDccAddress()).OrderBy(a => a).ToArray();
         for (var i = 0; i < orderedAdresses.Length; i++)
         {
             if (i < orderedAdresses.Length - 1 && orderedAdresses[i] + 1 == orderedAdresses[i + 1])
@@ -43,9 +56,9 @@ public static class LocoAddressExtensions
                 if (i < orderedAdresses.Length - 1) result.Append(',');
 
             }
-            //intervalStartIndex = -1;
         }
         return result.ToString();
     }
+    private static bool IsValidDccAddress(this int address) => address >= 1 && address <= 9999;
 
 }